structure refactor
