+++ hive-hooks/src/main/java/com/expediagroup/apiary/extensions/hooks/pathconversion/models/PathConversion.java
