+++ hive-hooks/src/test/resources/log4j2.xml
